# Reduce allocations in Resolver.Resolve.