--- a/rr.go
+++ b/rr.go
@@ -13,7 +13,7 @@ type RR struct {
 	Value string
 }
 
-type RRs []*RR
+type RRs []RR
 
 // emptyRRs is an empty, non-nil slice of RRs.
 // It is used to save allocations at runtime.
@@ -32,20 +32,28 @@ func (rr *RR) String() string {
 	return rr.Name + "\t      3600\tIN\t" + rr.Type + "\t" + rr.Value
 }
 
-func convertRR(drr dns.RR) *RR {
-	switch t := drr.(type) {
-	case *dns.NS:
-		return &RR{toLowerFQDN(t.Hdr.Name), dns.TypeToString[t.Hdr.Rrtype], toLowerFQDN(t.Ns)}
-	case *dns.CNAME:
-		return &RR{toLowerFQDN(t.Hdr.Name), dns.TypeToString[t.Hdr.Rrtype], toLowerFQDN(t.Target)}
-	case *dns.A:
-		return &RR{toLowerFQDN(t.Hdr.Name), dns.TypeToString[t.Hdr.Rrtype], t.A.String()}
-	case *dns.AAAA:
-		return &RR{toLowerFQDN(t.Hdr.Name), dns.TypeToString[t.Hdr.Rrtype], t.AAAA.String()}
-	case *dns.TXT:
-		return &RR{toLowerFQDN(t.Hdr.Name), dns.TypeToString[t.Hdr.Rrtype], strings.Join(t.Txt, "\t")}
-	default:
-		// fmt.Printf("%s\n", drr.String())
+// convertRR converts a dns.RR to an RR.
+func convertRR(drr dns.RR) (RR, bool) {
+	h := drr.Header()
+	rr := RR{
+		Name: toLowerFQDN(h.Name),
+		Type: dns.TypeToString[h.Rrtype],
 	}
-	return nil
+	switch t := drr.(type) {
+	// case *dns.SOA:
+	// 	rr.Value = toLowerFQDN(t.String())
+	case *dns.NS:
+		rr.Value = toLowerFQDN(t.Ns)
+	case *dns.CNAME:
+		rr.Value = toLowerFQDN(t.Target)
+	case *dns.A:
+		rr.Value = t.A.String()
+	case *dns.AAAA:
+		rr.Value = t.AAAA.String()
+	case *dns.TXT:
+		rr.Value = strings.Join(t.Txt, "\t")
+	default:
+		return rr, false
+	}
+	return rr, true
 }