Result:
-60% fewer allocs (30 → 12)
+45% more memory (1.3K → 1.8K)
+11% execution time (6.5µs → 7µs)

After:
BenchmarkResolve	  200000	      7218 ns/op	    1840 B/op	      12 allocs/op
BenchmarkResolveErr	  200000	      7151 ns/op	    1840 B/op	      12 allocs/op

Before:
BenchmarkResolve	  200000	      6500 ns/op	    1264 B/op	      30 allocs/op
BenchmarkResolveErr	  200000	      6447 ns/op	    1264 B/op	      30 allocs/op